Email password reset token to user upon password reset request
The requester receives a usable unhashed token link (one time usage with
10 min expiry) by email.
The token then gets hashed and stored in the DB to prevent others using
the real token.
Confirmed that using an unused but valid token correctly verifies against
the DB and then marks the token as used.
Also confirmed that verification fails if the stored token has expired.